Enum support for legacy fx
