+++ source/Sylvan.Data.Csv/Sylvan.Data.Csv.csproj
@@ -2,9 +2,9 @@
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0;netstandard2.1;netcoreapp3.0;net6.0</TargetFrameworks>
-    <VersionPrefix>1.1.1</VersionPrefix>
+    <VersionPrefix>1.1.2</VersionPrefix>
     <Description>A .NET library for reading and writing delimited CSV data.</Description>
-    <PackageTags>csv;delimited;datareader</PackageTags>
+    <PackageTags>csv;delimited;data;datareader;datawriter;simd</PackageTags>
     <Nullable>enable</Nullable>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <DefineConstants Condition="$(TargetFramework) == 'netstandard2.1'">SPAN;ASYNC_DISPOSE;$(DefineConstants)</DefineConstants>
